refactor : 로또 클래스 위치 원상복구
