+++ src/lotto/purchased-lottos.js
@@ -1,5 +1,5 @@
 import { Random, Console } from '@woowacourse/mission-utils';
-import Lotto from './Lotto.js';
+import Lotto from '../Lotto.js';
 import { LOTTO_RULES } from '../constant/index.js';
 
 class PurchasedLottos {
